utils: move listify to a generic trappy.utils file
listify() is used in both plotter and stats.  Move it out of
trappy.plotter.Utils into a generic trappy.utils file as it's not
something specific of the plotter.
