+++ trappy/plotter/AbstractDataPlotter.py
@@ -16,7 +16,7 @@
 """This is the template class that all Plotters inherit"""
 from abc import abstractmethod, ABCMeta
 from pandas import DataFrame
-from trappy.plotter.Utils import listify
+from trappy.utils import listify
 from functools import reduce
 # pylint: disable=R0921
 # pylint: disable=R0903
